fix(HeaderSection): Have serilization follow CoolSerial byte specs
It seems that cista does not do packed bytes for translating between structs.
This results in 6 bytes instead of 4 for the header section. Although cista
is likely doing checks to ensure accuracy; however, CoolSerial already
has its built-in verification.

There can be ways to increase performance; however, profiling will be slated
to see if the performance is justified in the first place.